requirements.txt
Flask==1.1.1
Flask-Login==0.4.1
pyOpenSSL==19.0.0
requests==2.22.0
google-cloud-datastore==1.9.0
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-oauthlib==0.4.0
google-auth-httplib2==0.0.3
flask-blueprint==1.3.0
python-rapidjson==0.8.0
gunicorn==19.9.0
pytz==2019.2
